Split deployment scripts
This splits these scripts out from what was a very long if/else block,
and implements some (potential) improvements in each. Both now log the
variables they will be using, and check whether those match what is
required a bit differently. The Konflux script also now uses `oc wait`
to check on some resource status details that it was previously polling:
this may require new-ish versions of `oc`, and was based on the v4.17
client.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>